Fixed a bunch of bugs:
Merged hse2states and states2prs because the state space is exactly the same as HSE and there was too much information to pass between programs.
Corrected the effective state calculation.
fixed a bug in generate_conflicts where I had the wrong conditions
Fixed a bug in bubble reshuffling where it couldn't recognize a cycle and exit with an error
Fixed a few bugs surrounding the reset generation